backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0
djangorestframework-simplejwt>=5.2.0
mysqlclient>=2.1.0
python-dotenv>=1.0.0
Pillow>=9.0.0
django-cors-headers>=4.0.0
django-filter>=22.1
django-extensions
